requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2-binary
celery[redis]==5.2.7
flower==1.2.0
celery_singleton==0.3.1
